add data prep and training scripts
Bifurcate the scripts to isolate the different aspects, i.e. create an external location in databricks to the clean dataset container in adlsg2, prep_dataset to split and crop the dataset to adhere to the model training parameters and finally train_model to set up experiments, train the model and log metrics.